requirements.txt
fastapi
uvicorn
python-dotenv
edge-tts
mutagen
pydantic
groq
moviepy